requirements.txt
altair==4.2.2
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
entrypoints==0.4
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.20.1
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
Sastrawi==1.0.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
streamlit-aggrid==1.0.5
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
xgboost==2.1.3
